Fix theme flickering/flashing during pages loading
When an user has opposite themes for their OS and the website, like OS
dark and website light, it generates a flickering behaviour during
loading pages while browsing the website.

This fix this bug introducing a single line js inside the <head> to update the
"data-theme" if there is a local "theme""defined by the user.

This addresses issue #429.